Updated client server

+++ include/tcp_sock_handler.h
@@ -11,10 +11,10 @@ struct TCP_SOCKET_HANDLER
 typedef struct TCP_SOCKET_HANDLER *TCP_HANDLER;
 
 TCP_HANDLER new_tcp_handler(int socket_fd);
-size_t tcp_sendn(TCP_HANDLER handler, char *buf, size_t buf_len);
-size_t tcp_recvn(TCP_HANDLER handler, char *buf, size_t buf_len);
-size_t tcp_send(TCP_HANDLER handler, char *buf, size_t buf_len);
-size_t tcp_recv(TCP_HANDLER handler, char **buf, size_t buf_len);
-size_t destroy_tcp_handler(TCP_HANDLER handler);
+int tcp_sendn(TCP_HANDLER handler, char *buf, int buf_len);
+int tcp_recvn(TCP_HANDLER handler, char *buf, int buf_len);
+int tcp_send(TCP_HANDLER handler, char *buf, int buf_len);
+int tcp_recv(TCP_HANDLER handler, char **buf, int buf_len);
+int destroy_tcp_handler(TCP_HANDLER handler);
 
 #endif
